Use mixin method for clearing enrichments
