+++ server/src/main/java/io/spine/server/aggregate/AggregateStorage.java
@@ -35,7 +35,6 @@ import java.util.Optional;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.protobuf.util.Timestamps.checkValid;
-import static io.spine.core.Events.clearEnrichments;
 import static io.spine.validate.Validate.checkNotEmptyOrBlank;
 
 /**
@@ -105,7 +104,7 @@ public abstract class AggregateStorage<I>
     /**
      * Writes an event to the storage by an aggregate ID.
      *
-     * <p>Before the storing, {@linkplain io.spine.core.Events#clearEnrichments(Event) enrichments}
+     * <p>Before the storing, {@linkplain io.spine.core.Event#clearEnrichments() enrichments}
      * will be removed from the event.
      *
      * @param id    the aggregate ID
@@ -114,7 +113,7 @@ public abstract class AggregateStorage<I>
     void writeEvent(I id, Event event) {
         checkNotClosedAndArguments(id, event);
 
-        Event eventWithoutEnrichments = clearEnrichments(event);
+        Event eventWithoutEnrichments = event.clearEnrichments();
         AggregateEventRecord record = toStorageRecord(eventWithoutEnrichments);
         writeRecord(id, record);
     }
